examples/streamlit-docker-openai/requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0

